proxy/requirements.txt
# requirements.txt
flask==3.0.0
openai==1.3.0
flask-limiter==3.5.0
redis==5.0.1
python-dotenv==1.0.0
flask-cors==4.0.0